pyproject.toml
[project]
name = "sp-index-lab"
version = "0.1.0"
description = "AI-powered portfolio analytics platform proving the S&P 500 is driven by ~20 stocks"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30",
    "plotly>=5.18",
    "pandas>=2.1",
    "numpy>=1.26",
    "yfinance>=0.2.31",
    "supabase>=2.0",
    "duckdb>=0.9",
    "pyportfolioopt>=1.5",
    "scikit-learn>=1.3",
    "lightgbm>=4.1",
    "statsmodels>=0.14",
    "vectorbt>=0.26",
    "hmmlearn>=0.3",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
